requirements.txt
asgiref==3.5.2
Django==4.1.3
django-js-asset==2.0.0
django-mptt==0.14.0
factory-boy==3.2.1
Faker==15.3.3
psycopg2==2.9.5
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
